search bar no longer refreshes entire header
