+++ client/src/index.css
@@ -57,7 +57,7 @@ footer {
 
 
 main {
-  height: 80vh;
+  height: 90vh;
   display: flex;
   flex-direction: column;
   justify-content: center;
